backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.31
alembic==1.13.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
stripe==9.12.0
pydantic[email]==2.7.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
